chore(ops): Move check ops code
