+++ src/synapse/__init__.py
@@ -61,9 +61,6 @@ from .workload import (  # noqa: F401
     SYNAPSE_USER,
     ExecResult,
     WorkloadError,
-    check_irc_bridge_ready,
-    check_mjolnir_ready,
-    check_nginx_ready,
     create_irc_bridge_app_registration,
     create_irc_bridge_config,
     create_mjolnir_config,
